- bug fix: if has Reimbursement Guide Entries, then must not be deleted; furthermore, Reimbursement Guide Entries are not to be deleted - added public canBeDeleted method - bug fix: corresponding Guide's total value must be updated

+++ src/net/sourceforge/fenixedu/domain/GuideEntry.java
@@ -1,5 +1,7 @@
 package net.sourceforge.fenixedu.domain;
 
+import net.sourceforge.fenixedu.domain.exceptions.DomainException;
+
 
 
 /**
@@ -26,15 +28,25 @@ public class GuideEntry extends GuideEntry_Base {
     }
 
     public void delete() {
-        if (hasPaymentTransaction()) {
-            getPaymentTransaction().delete();
-        }
-        for (; hasAnyReimbursementGuideEntries(); getReimbursementGuideEntries().get(0).delete());
-        
-        removeGuide();
-        removeRootDomainObject();
+        if (canBeDeleted()) {
+            if (hasPaymentTransaction()) {
+                getPaymentTransaction().delete();
+            }
+            
+            Guide guide = getGuide();
+            removeGuide();
+            guide.updateTotalValue();
+            
+            removeRootDomainObject();
 
-        deleteDomainObject();
+            deleteDomainObject();
+        } else {
+            throw new DomainException("guide.entry.cannot.be.deleted");
+        }
     }
 
+    public boolean canBeDeleted() {
+        return !hasAnyReimbursementGuideEntries();
+    }
+    
 }
